K Mean implemented. Weird Results
I implemented K means using sparse function representations

The results are very weird though.

For k in [1, 2, 3] and various random initial cluster representatives
I've found that all representatives end up the same.

Yep
So if you run k-means with k = 3, after 1 iteration all the
representatives are the same, which is equal to the mean of the
reconciliation graph.

Either there's a bug or this is a really weird space.