Reconciliation Finding + KMeans, Testing
I've spent a lot of time testing. In particular testgen now holds two
functions which are useful for constructing simple 'reconciliation
graphs' which are easy to reason about (and we can thus verify the
algorithm is behaving correctly). The quotation marks indicate that
these are not graphs which encode real reconciliations.

I've also hooked up the reconciliation graph constructing code and the
clustering code. The result in `FromNewick.py`